Added other plugin metadata
PlotMe will now allow other plugins to store plot properties or metadata
into plots. This data will be kept in the plotme database. If a plot is
reset, the metadata is will be removed. If a plot is moved, the metadata
will follow to the new plot.

See the PlotMeHookExample plugin for an example usage
https://github.com/WorldCretornica/PlotMeHookExample

Fixes :
* Removed redundant code from the plot move function
* Fixed code in the SQL addPlot function that would try to fetch empty
string from mojang
* Removed logger from the UUID fetcher for now to keep the API
bukkit-free.